Fixed C permutation

+++ 2.n-puzzle/solver/src/permutation.h
@@ -1,1 +1,1 @@
-long long pnum(int n, int k, int *a);
+long long pnum(long long n, long long k, long long *a);
